refactor SeoLinker extension to controller structure
