+++ hugashop/crm/Extensions/SeoLinker/settings.yaml
@@ -3,4 +3,4 @@ description: "Анализ внутренних ссылок сайта"
 settings_params:
   - { variable: base_url, name: "Стартовая страница сканирования" }
   - { variable: delay, name: "Задержка сканирования, мсек.", default: 0 }
-version: 1.2
+version: 1.3
